services/video-worker/requirements.txt
boto3==1.28.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pytest==7.4.0

